More work.

+++ src/main/java/org/halvors/nuclearphysics/client/gui/component/IBarInfoHandler.java
@@ -5,7 +5,7 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 
 @SideOnly(Side.CLIENT)
 public interface IBarInfoHandler {
-    int getLevel();
+    double getLevel();
 
     default String getTooltip() {
         return null;
